+++ packages/next-swc/crates/next-dev-tests/tests/integration/next/app/route/input/app/[slug]/route.ts
@@ -21,3 +21,5 @@ export const POST = helloHandler
 export const PUT = helloHandler
 export const DELETE = helloHandler
 export const PATCH = helloHandler
+
+export const runtime = "experimental-edge";
